refactor(db): extract checkpoint coordinator state
